requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
opencv-python
mss
numpy
openai
python-dotenv
pyaudio
SpeechRecognition
soundfile